+++ hmr-test/main.js
@@ -1,0 +1,3 @@
+import "./styles.css";
+import { Elm } from "./src/App.elm";
+Elm.App.init();